refactor: 댓글 더보기 버튼 내부로직 사용자 훅으로 분리
