+++ src/components/Element/Buttons/ButtonOptionComment.jsx
@@ -1,39 +1,17 @@
-import React, { useState } from 'react';
+import React from 'react';
 import './style/OtherButton.scss';
-import { useLocation } from 'react-router-dom';
-import toast from 'react-hot-toast';
 import Modal from '../Modal/Modal';
-import { deleteCommentAPI, postCommentReportAPI } from './api';
+import useHandleCmtOptionButton from '../../../hooks/useHandleCmtOptionButton';
 
 function ButtonOptionComment({ commentid, isMyCmt }) {
-  const [isModal, setIsModal] = useState(false);
-  const location = useLocation();
-  const postid = new URLSearchParams(location.search).get('postId');
-
-  const handleModal = () => {
-    setIsModal(prev => !prev);
-  };
-
-  const handleDelete = async () => {
-    const res = await deleteCommentAPI(postid, commentid);
-    if (res.status === '200') window.location.reload();
-  };
-
-  const handleReport = async () => {
-    const res = await postCommentReportAPI(postid, commentid);
-    if (res.report) toast('댓글 신고 완료', { icon: '⚠️', position: 'bottom-center', ariaProps: { role: 'alert', 'aria-live': 'polite' } });
-  };
-
-  const myOptions = [{ text: '삭제', func: handleDelete, openAlert: true }];
-
-  const options = [{ text: '신고', func: handleReport, openAlert: true }];
+  const { isModal, handleModal, options } = useHandleCmtOptionButton(commentid, isMyCmt);
 
   return (
     <>
       <button type="button" className="btn-option" onClick={() => handleModal()}>
         <span className="a11y-hidden">더보기</span>
       </button>
-      {isModal && <Modal options={isMyCmt ? myOptions : options} closeModal={handleModal} />}
+      {isModal && <Modal options={options} closeModal={handleModal} />}
     </>
   );
 }
